refactor: move packages for better structure
