+++ src/main/java/net/simplyvanilla/simplyrank/SimplyRankPlugin.java
@@ -6,14 +6,14 @@ import com.google.gson.GsonBuilder;
 import net.kyori.adventure.text.format.NamedTextColor;
 import net.kyori.adventure.text.format.TextColor;
 import net.simplyvanilla.simplyrank.command.SimplyRankCommandExecutor;
-import net.simplyvanilla.simplyrank.data.GroupPermissionService;
-import net.simplyvanilla.simplyrank.data.PermissionApplyService;
-import net.simplyvanilla.simplyrank.data.PlayerDataService;
-import net.simplyvanilla.simplyrank.data.PlayerPermissionService;
-import net.simplyvanilla.simplyrank.data.database.group.GroupData;
-import net.simplyvanilla.simplyrank.data.database.sql.MySqlClient;
-import net.simplyvanilla.simplyrank.data.database.sql.MySqlRepository;
-import net.simplyvanilla.simplyrank.exception.DatabaseConnectionFailException;
+import net.simplyvanilla.simplyrank.permission.GroupPermissionService;
+import net.simplyvanilla.simplyrank.permission.PermissionApplyService;
+import net.simplyvanilla.simplyrank.permission.PlayerDataService;
+import net.simplyvanilla.simplyrank.permission.PlayerPermissionService;
+import net.simplyvanilla.simplyrank.database.group.GroupData;
+import net.simplyvanilla.simplyrank.database.sql.MySqlClient;
+import net.simplyvanilla.simplyrank.database.sql.MySqlRepository;
+import net.simplyvanilla.simplyrank.database.exception.DatabaseConnectionFailException;
 import net.simplyvanilla.simplyrank.gson.TextColorGsonDeserializer;
 import net.simplyvanilla.simplyrank.listener.PlayerLoginEventListener;
 import net.simplyvanilla.simplyrank.listener.PlayerQuitEventListener;
