+++ src/main/java/net/simplyvanilla/simplyrank/database/sql/exceptions/MySqlStatementFailedException.java
@@ -1,4 +1,4 @@
-package net.simplyvanilla.simplyrank.data.database.sql.exceptions;
+package net.simplyvanilla.simplyrank.database.sql.exceptions;
 
 public class MySqlStatementFailedException extends RuntimeException{
     public MySqlStatementFailedException(Throwable cause) {
