+++ src/main/java/net/simplyvanilla/simplyrank/listener/PlayerQuitEventListener.java
@@ -1,6 +1,6 @@
 package net.simplyvanilla.simplyrank.listener;
 
-import net.simplyvanilla.simplyrank.data.PlayerPermissionService;
+import net.simplyvanilla.simplyrank.permission.PlayerPermissionService;
 import org.bukkit.event.EventHandler;
 import org.bukkit.event.Listener;
 import org.bukkit.event.player.PlayerQuitEvent;
