+++ src/test/java/net/simplyvanilla/simplyrank/PermissionApplyServiceTest.java
@@ -6,14 +6,14 @@ import be.seeseemelk.mockbukkit.ServerMock;
 import be.seeseemelk.mockbukkit.UnimplementedOperationException;
 import be.seeseemelk.mockbukkit.entity.PlayerMock;
 import net.kyori.adventure.text.format.NamedTextColor;
-import net.simplyvanilla.simplyrank.data.GroupPermissionService;
-import net.simplyvanilla.simplyrank.data.PermissionApplyService;
-import net.simplyvanilla.simplyrank.data.PlayerDataService;
-import net.simplyvanilla.simplyrank.data.PlayerPermissionService;
-import net.simplyvanilla.simplyrank.data.database.group.GroupData;
-import net.simplyvanilla.simplyrank.data.database.group.GroupRepository;
-import net.simplyvanilla.simplyrank.data.database.player.PlayerData;
-import net.simplyvanilla.simplyrank.data.database.player.PlayerDataRepository;
+import net.simplyvanilla.simplyrank.permission.GroupPermissionService;
+import net.simplyvanilla.simplyrank.permission.PermissionApplyService;
+import net.simplyvanilla.simplyrank.permission.PlayerDataService;
+import net.simplyvanilla.simplyrank.permission.PlayerPermissionService;
+import net.simplyvanilla.simplyrank.database.group.GroupData;
+import net.simplyvanilla.simplyrank.database.group.GroupRepository;
+import net.simplyvanilla.simplyrank.database.player.PlayerData;
+import net.simplyvanilla.simplyrank.database.player.PlayerDataRepository;
 import net.simplyvanilla.simplyrank.database.GroupRepositoryMock;
 import net.simplyvanilla.simplyrank.database.PlayerDataRepositoryMock;
 import net.simplyvanilla.simplyrank.listener.PlayerLoginEventListener;
